generate custom playlist backend stuff

+++ src/main/java/edu/brown/cs/cjps/music/PlaylistDefaults.java
@@ -15,7 +15,7 @@ public class PlaylistDefaults {
   public Settings getWorkStudyDefaults() {
     List<String> g = Arrays.asList("classical", "jazz");
     float m = 0.8f;
-    float hotness = 40f;
+    int hotness = 40;
     float energy = 0.3f;
     return new Settings(Tag.WORKSTUDY, g, m, hotness, energy);
   }
@@ -23,14 +23,14 @@ public class PlaylistDefaults {
   public Settings getEatSocialDefaults() {
     List<String> g = Arrays.asList("pop", "rock", "folk");
     float m = 0.6f;
-    float hotness = 60f;
+    int hotness = 60;
     float energy = 0.4f;
     return new Settings(Tag.EATSOCIAL, g, m, hotness, energy);
   }
 
   public Settings getExerciseDefaults() {
     List<String> g = Arrays.asList("hip-hop", "rock");
-    float hotness = 70f;
+    int hotness = 70;
     float energy = 1.0f;
     float m = 0.5f;
     return new Settings(Tag.EXERCISE, g, m, hotness, energy);
@@ -38,7 +38,7 @@ public class PlaylistDefaults {
 
   public Settings getPartyDefaults() {
     List<String> g = Arrays.asList("pop", "hip-hop");
-    float hotness = 100f;
+    int hotness = 100;
     float energy = 1.0f;
     float m = 0.8f;
     return new Settings(Tag.PARTY, g, m, hotness, energy);
@@ -46,7 +46,7 @@ public class PlaylistDefaults {
 
   public Settings getRestfulDefaults() {
     List<String> g = Arrays.asList("classical");
-    float hotness = 30f;
+    int hotness = 30;
     float energy = 0.2f;
     float m = 0.5f;
     return new Settings(Tag.RESTFUL, g, m, hotness, energy);
